Fixed application to work with docker
The application now uses environment variables to identify where the
backend is instead of the previously hardcoded "localhost".

To use this, when running the application ensure to set the API_URL
environment variable to point to the backend address and port.

Example(Running backend on my local network):
$ export API_URL="192.168.0.232:8080"
$ docker-compose up

Forgetting to define API_URL will lead to errors